member_menu_items moved to users_helper
